update to have 2 manifest and markdown scripts
